Fix the problem of compilation on Windows in another way: use decimal index (long long), not real (long double) to output it to string without ".00"

+++ src/GeomAlgoAPI/GeomAlgoAPI_MakeShape.cpp
@@ -209,7 +209,7 @@ void GeomAlgoAPI_MakeShape::initialize() {
 //=================================================================================================
 void GeomAlgoAPI_MakeShape::prepareNamingFaces()
 {
-  long double index = 1;
+  long long index = 1;
   GeomAPI_ShapeExplorer anExp(shape(), GeomAPI_Shape::FACE);
   for(GeomAPI_ShapeExplorer anExp(shape(), GeomAPI_Shape::FACE); anExp.more(); anExp.next()) {
     std::shared_ptr<GeomAPI_Shape> aFace = anExp.current();
